Listagem e exclusão do histórico de Export

+++ app/Http/Controllers/ExportController.php
@@ -5,33 +5,37 @@ namespace App\Http\Controllers;
 use App\Models\Export;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Inertia\Inertia;
 
 class ExportController extends Controller
 {
-    /**
-     * This function is responsible for handling the index action of the ExportController.
-     * It does not perform any specific operations but can be used to display a list of exports or any other relevant data.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
-        $exports = Export::paginate(15);
+        $exports = Export::paginate(5);
+
+        return Inertia::render('Reports', [
+            'exports' => $exports
+        ]);
     }
 
-    /**
-     * Delete the specified export from storage.
-     *
-     * @param  \App\Models\Export  $export
-     * @return \Illuminate\Http\Response
-     *
-     * @throws \Exception
-     */
-    public function destroy(Export $export)
-    {
-        Storage::delete($export->file_name);
-        $export->delete();
+    public function show($export) {
+        $export = Export::find($export);
 
-        return 'Delete';
+        return Storage::download($export->file_name);
+    }
+
+    public function destroy($export)
+    {
+        $export = Export::find($export);
+
+        if ($export) {
+            Storage::delete($export->file_name);
+            $export->delete();
+        }
+
+        return redirect()->back()
+            ->with('success', 'Seu arquivo foi removido com sucesso');
+
     }
 }
